Update compiler and syntax

+++ src/util/utility/util_display.h
@@ -1,5 +1,5 @@
 ///////////////////////////////////////////////////////////////////////////////
-//  Copyright Christopher Kormanyos 2019 - 2023.
+//  Copyright Christopher Kormanyos 2019 - 2024.
 //  Distributed under the Boost Software License,
 //  Version 1.0. (See accompanying file LICENSE_1_0.txt
 //  or copy at http://www.boost.org/LICENSE_1_0.txt)
@@ -23,7 +23,7 @@
     virtual auto init() noexcept -> bool = 0;
 
     virtual auto write(const char* pstr,
-                       const std::uint_fast8_t length,
+                       const std::size_t length,
                        const std::uint_fast8_t line_index) noexcept -> bool = 0;
 
   protected:
